RakuAST: fix rules for determing implied code blocks
Previously, it was assumed that the indentation level of the
=begin pod would be taken as a base to determine an implied code
block.  But it turns out it is indentation level of the last seen
doc block.

=begin pod
    =head1 Foo

    This is *not* an implied code block
=end pod

This moves the indentation of abbreviated doc blocks into its
contents so that there is no difference between:

    =head1 Foo

and

    =begin head1
    Foo

with regards to contents.  This also eliminates the :spaces argument
to the RakuAST::Doc::Block.from-paragraphs method, as this is no longer
needed.