RakuAST: implement DOC phaser support
For now, this works as intended when --doc is specified.  It dies
rather spectactularly when --doc is not specified: need to find a
way to disable the <blorst> from being processed in that case.

With regards to DOC CHECK, that just implements the CHECK phaser,
which currently has INIT semantics.  When that is fixed, this will
be fixed as well.