RakuAST: add .literalize method to selected ::Node classes
RakuDoc configuration can become very complex, basically everything
is allowed, as long as the result is a constant "literal" value if
you will.

This adds a src/core.c/Literalize.pm6 file that augments the
RakuAST::Node class with "literalize" methods for the RakuAST::Node
classes that we've seen used in Rakudoc config tests.

Adapt the actions to use the "literalize" method on the ast of the
config.  If Nil is returned, then the config is *not* constant.
Otherwise the HLL representation of the config is returned.

It seems likely that this functionality might actually also be used
for some constant folding purposes, by running it on an AST and
codegen a WVal of the returned value if it is not Nil.