RakuAST: ensure that pod blocks are generated in right order
By giving each RakuAST::Doc block that will be generated into Pod::
objects an ordinal number at the moment they're created, and use that
number as in index into $=pod.

Sadly, not all created RakuAST::Doc blocks get the CHECK handling
(not entirely know why), so holes are created in $=pod.  So at the
end of CHECK time, make sure any holes are removed from $=pod.

If and/or when $=pod no longer needs to be generated, can we remove
this hack.