RakuAST: make sure Subset.WHY is correct with a where block
Because bare blocks can also have declarator docs attached, and
a "where" block of a subset is seen *before* the subset, the
declarator doc was attached to the block of the subset, rather
than to the subset itself.

This introduces a "cut-WHY" method on declarator targets, and
a steal-declarand method in Actions, and the use of these when
appropriate in the subset handling.