Deprecate including an ambiguous .pm module file
We need to know when raku source files have changed for the purposes
of precompilation. However, when a user includes a library path
(i.e. -Ilib, -I/home/foo) that does not contain a META6.json file
we might end up erroneously considering perl module files (which
also use .pm) as raku files. While this is mostly harmless, it
does waste some cpu and io. This adds a deprecation message when
this ambiguous situation occurs so that in the future we might be
able to remove support for the .pm extension altogether.