Make extensions for CURFS configurable
When using CURFS that isn't pointed at a directory containing a
META6.json file (i.e. -Ilib) it will consider .rakumod .pm6 and
.pm files as raku modules. However, some users might want to
opt out of e.g. .pm files getting treated in this way.

This adds the ability for users to define which extensions a CURFS
will scan for using the CompUnit::Repository::Spec format. For
example: file#extensions<.rakumod .pm6>#lib would only scan for
files with the .rakumod and .pm6 extension of the lib directory.