RakuAST: fix BEGIN time CORE:: lookup
At BEGIN time, we don't have the setting as an outer frame for
PseudoStash to find. The old frontend got around this by creating a
jumbo frame containing all lexically visible symbols, which just so
happened to include CORE-SETTING-REV which is what PseudoStash was
looking for.

Instead, codegen CORE:: differently during BEGIN time compilation.
We can create an appropriately prepared PseudoStash object at compile
time and just insert it into the QAST as a WVal. We cannot do that in
normal compilation as we can't serialize the setting context. But these
dynamically compiled frames won't be serialized anyway, so we should be
good.