Remove stat existence check
Currently we do a stat to check that a path exists before we do
a followup stat to check if that path is a file or directory. We
have to do the second stat regardless, so if we just 'try' the
second stat and remove the existence check then we ultimately
do one less stat call. This changes the current behavior to the
later aforementioned behavior, resulting in a ~10% speed up.