Filter paths before stating them
Currently the default of DIR-RECURSE ignores any directory that
starts with '.'. However, when reading the contents of a directory
there is always a '.' and '..', which means we are stat'ing a lot
of $entry that get filted out in the next conditional i.e.
$!dir.ACCEPTS. This reorders the file and directory check
conditionals to filter before doing a stat, resulting in a ~3%
speed up with the default directory and file filters.