Reduce number of 'try' used
Previously we added 'try' to nqp ops that expect their path arg
to exist on the file system in order to remove a stat call to
check for existence. This reduces the number of 'try' used from
one for each of the aforementioned nqp ops to just one total,
which results in a ~3% speed up.