Respect CUPR.may-precomp() in CURI
CompUnit::PrecompilationRepository.may-precomp() can be used
to check if something should try to precompile using said repo.
It is used in CUPR in its try-load method to determine if it should
try to precompile a file when loading, which often happens with
a CURFS type repository. However, CURI.install also tries to
precompile files, but it does not check may-precomp() before doing
so. This means there is no central place to enable or disable
precompiling.

This updates CURI to only precompile modules during install if
may-precomp() is True.