Add Format.AST
Since Format isa Str, it already had an .AST method, but that would
try to interpret the format string as Raku code, rather than as a
sprintf format.  By adding this method, at least any confusion in
that area is averted.

As a bonus, since the .AST returns a RakuAST node, you can call
.DEPARSE on that, so that you can more easily see what code was
actually generated for a format string, e.g.

    $ RAKUDO_RAKUAST=1 raku -e 'use v6.*; say q:o/%s/.AST.DEPARSE'
    -> $a {
        $a.Str()
    }