Make sub version of min/max/minmax on Ranges consistent
with the method equivalent.  Since "min" and "max" are attributes
on the Range object, they will *always* reflect the "left" and "right"
values in the Range (even for Ranges such as (10..-10).

There were no sub versions of min/max/minmax for Ranges, so they
followed Iterable logic.  Which would give different results from
the method version.  And rather unexpected ones as well:

    say max 10..-10;  # -Inf, because (10..-10).list is an empty List
    say max ^3;       # 2, because (^3).list is (0,1,2)