Reorganize Format a bit
While documenting the Format class, it occurred to me that the "code"
method was pretty ambiguous in its naming.  Replaced that by the
"Callable" method.  Also removed the .signature method, as that now
more logically falls out of Callable.signature.  Kept the "arity"
and "count" methods as these are deemed to be more necessary for
properly feeding the Callable from different sources (e.g. the .fmt
method).