Copy modules from lib/ to individual distributions at core-lib/
As the first step of splitting the core modules into multiple
distributions we will copy the modules into subdirectories of
core-lib/, grouping them in a way that seems logical. In a
future step we will incorporate these new paths to various parts
of the tooling (since we will need to do -Icore-libs/BUILDPLAN,
-Icore-libs/NativeCall, and so on) and delete the still existing
lib/ directory.