RakuAST: a better way to extract meta-info from markup
Take the last atom: it *must* be a Str because we don't allow markup
in the meta of L<> and X<>.  Split that on the first | seen (which
could well be the first character).  Use that as meta.  Adjust atom
accordingly, or remove it if it has now become empty.

This fixes some corner cases in the Raku documentation, but breaks
other cases, such as L<foo::(bar|baz)|url>, which will now have to
be written as L<V<foo::(bar|baz)>|url>, which sadly isn't rendered
correctly by Pod::To::Text.