RakuAST: handle synthetics in Markup parser properly
The initial implementation assumed that one could use the same index
in both the Uni representation, as well as an index in nqp::eqat.
Synthetic codepoints break this assumption, as they are always decoded
as decomposed.

So change the stopper logic into a single stopper value to be checked
and if that matches, check if a multi grapheme stopper was involved
(e.g. in the case of << >>).

This change may actually speed up parsing, as now only a single check
is done for each graphame to determine stopper, instead of two.