Remove rakudo-home/lib from vm_search_path
rakudo-home/lib is provided elsewhere, and thus this search path
isn't needed. This makes the vm_search_path return an empty array,
similar to the JS backend. Note that we don't get rid of the
vm_search_path method altogether because the JVM can return a non
empty array of paths.