Add support for Version.new("1,2,3-")
This adds support for Version objects that will accept any Version
equal or **before** a given Version object.

    say Version.new("1.2.3") ~~ Version.new("1.2.4-");  # True
    say Version.new("1.2.4") ~~ Version.new("1.2.4-");  # True
    say Version.new("1.2.5") ~~ Version.new("1.2.4-");  # False

    say Version.new("1.2.3") cmp Version.new("1.2.4-");  # Less
    say Version.new("1.2.4") cmp Version.new("1.2.4-");  # More
    say Version.new("1.2.5") cmp Version.new("1.2.4-");  # More

As discussed in:
  https://github.com/finanalyst/Rakudoc-BETA/issues/5#issuecomment-1620824362

Implemented internally by giving an additional meaning to the $!plus
attribute: when it is 1, it indicates "+", when it is -1 it indicates
"-" and when 0 it indicates absence of both "+" and "-".

Adapted .plus for these new mnemonics, and added a .minus method which
will return True if the Version had a "-" specified.

It is intended that this functionality will become available in
Version literal strings in RakuAST, and thus in 6.e in the future.
So no semantic changes for 6.c/d.