RakuAST: get rid of rakudoc-verbatim-blocks token
For some reason, this caused a regression in doc/Language/tables.rakudoc
parsing, related, golfed down to:

    =begin code
    =for table
            mouse    | mice
    =end code

which caused the table to be parsed as a table, instead of verbatim
text.  No idea why.  But since it was only being used in one location
so far, this appears to be the easiest fix and way of least resistance.