Implement support for role's COMPOSE submethod in Raku v6.e
The submethod is implementing the funcionality many expect from a phaser
of the same name. But the submethod can be considered more fit to the
task as it's kind of getting into the line with construction submethods
`BUILD` and `TWEAK` with the difference that it is role-specific and it
is about type-composition time, not object-composition.

The submethod is invoked on role concretization before it gets applied
to a class.