Make c3_merge method independent of C3MRO role
As a matter of fact, it would better do as a sub, but we don't use
global subs in the Metamodel. Therefore, for now, I move it into
Perl6::Metamodel::Configuration. This is a temporary experimental
solution as it would make more sense to have a class dedicated to
utility methods of similar kind.