Implement methods `andthen` and `orelse` on Promise
The methods are intended to simplify conditional processing and chaining
of kept or broken promises when used instead of the `then` method.

`andthen` is fired only when its invocant promise is kept. If the
invocant is broken then its cause becomes the cause of the promise
returned by `andthen`:

```
my $p = Promise.new;
my $p2 =
    $p.andthen({ note "must not happen!" }) # will print nothing
      .then({ say "cause: ", ~.cause });    # will print "cause: broken"
$p.break("broken");
try await $p2;
```

`orelse` is fired if its invocant is broken. If the invocant is kept
then the promise returned by `orelse` is kept with invocant's result:

```
my $p = Promise.new;
my $p2 =
    $p.orelse({ note "this must not be printed!" })
      .then({ say "result: ", ~.result });
$p.keep("all good");
await $p2;
```

If the invocant of `orelse` is broken the purpose of the method is to
intercept the cause, handle it, and return a value to be used as the
result of `orelse`-returned promise:

```
my $p = Promise.new;
my $p2 =
    $p.orelse({ "purified" })
      .then({ say "result: ", ~.result });
$p.break("all bad");
await $p2;
```