Implement methods `andthen` and `orelse` on Promise
The methods are intended to simplify conditional processing and chaining
of kept or broken promises when used instead of the `then` method.

`andthen` is fired only when its invocant promise is kept. If the
invocant is broken then its cause becomes the cause of the promise
returned by `andthen`:

`orelse` is fired if its invocant is broken. If the invocant is kept
then the promise returned by `orelse` is kept with invocant's result:

If the invocant of `orelse` is broken the purpose of the method is to
intercept the cause, handle it, and return a value to be used as the
result of `orelse`-returned promise: