RakuAST: make :allow a purely renderer thing
This means that any RakuDoc (apart from =comment and =data) will
have its paragraphs checked for Markup codes, and generate the
associated Markup objects.  It is then up to the renderer to look
at any :allow setting (or absence of that) to determine whether
the Markup should be interpreted, or just be stringified.

The requirement for a valid Markup letter is now having "Lu" as
the general Unicode property.  So this will allow Δ<> markup.