RakuAST: abstract allowed markup determination
Added method ::Doc::Block.allowed-markup, which returns a Map that
conceptually interrogated with a letter to see if that letter is
allowed or not.  If there is a specific ":allow" active on the block,
then a Map with those letters will be returned.  If there is not,
then either a special Map that returns True for any letter that has
the "Lu" Unicode property, or one that always returns False for any
letter.  All depending on the type of the block.

Use this new functionality in the legacy podification.