Fix stringification of v6.*
Commit 4f07e0e1a9280595c9a7 removed the whatever attribute.  This
changed the stringification of v6.* to "v6.e.PREVIEW", which caused
test issues while installing the META6 distribution.

Fix this by creating a special case for v6.*, which will compare as
a v6.e.PREVIEW, but which will stringify as "v6.*".  This fixes the
test problem on the META6 distribution and is spectest clean.

    say v6.* cmp v6.e.PREVIEW;  # Same