RakuAST: add support for "also is Foo" and "also does Foo"
Make this a purely grammar / action event, without any explicit
deparsing or .raku.  While 7114b4d739 introduced a flag for Traits
to indicate they were from an "also", and while that would help in
deparsing, it would not help with .raku.

That's because when an `Statement::Also` statement would get executed,
it would either have to specifically specify the package to attach
the traits to, or depend on some environment variable, neither of
which would be giving a nice interface.  And if you would want to
programmatically add a Trait to a Package, then there's the .add-trait
method already.

So this means that:

    class A { also is Int }

will deparse as:

    class A is Int { }

and similarly for .raku.  Sometimes you can't have it all.