Don't use CALLERS:: to lookup %?LANG
I doubt it makes sense whatsoever. When there is a call to `EVAL` it'd
be rather expected to inherit the language of the immediate lexical
scope of the caller. I don't see why a scope of a distant caller from
the stack should be affecting this. Moreover, it is likely to cause some
sorts of problems with compilation if the EVALed code is not compatible
with that caller's slang.