RakuAST: channel op= through METAOP_ASSIGN
Just like with the legacy grammar.  This has the disadvantage of
needing to call a sub (which in turn calls a Rakudo::Internals
method) for each op= execution.

Another solution would be to directly codegen the meta-op for the
given infix op.  Sadly, this effort created all sorts of weird
errors that I have not been able to grok yet.

At least it doesn't warn executing an op= on an undefined left-hand
side anymore.