Fix issue with uint64 Bufs and -1
Apparently there's some kind of coercion going on when there's not
a direct assignment from a uint to an int.  This coercion is throwing
the error.

Fix this by making it direct assignments: for the represention in a
gist, it doesn't matter whether the value was originally signed or
unsigned.

Fixes #5348