RakuAST: Streamline lang-setup a bit
And also introduce the RAKU_LANGUAGE_VERSION environment variable
to set the language version to be used if no "use v6xxx" has been
seen.

Sadly this exposed a bug in language version setting: $*RAKU is
set correctly with e.g. a "use v6.*", but the actual grammar being
used, is still the default, aka the 6.d grammar.