Un-nqpify *.combinations() without slowing it down
This reduces the amount of `nqp::*` ops used in the implementation of
`*.combinations()`. In my timing of `say (1..100).combinations(4).tail;
say now - INIT now` this causes no noticeable slowdown, any further
reductions do. There are no logical changes, but it might make it easier
to experiment with alternate algorithms or optimizations.