Introduce Operator class, subclass of Sub
Historically, infix/prefix/postfix/postcircumfix operators are just
ordinary subs with a special name.  Certain aspects of operators just
lived inside hashes in the legacy grammar.

For the Raku grammar, it is the intent to get rid of these grammar
structures, as one would like to be able to define operators using
RakuAST.  And that means being able to assign these aspects to the
objects *without* the use of a grammar.

So the OperatorProperties class was created for this purpose.  And
an "op_props" attribute was added to the Routine class quite a while
ago already.  This was a mistake.

First of all, the attribute should have been added to the Sub class,
as operators can only be subs, not methods or submethods.

Secondly, these OperatorProperties only apply to a small subset of
subroutines.

So it felt to be a better approach to introduce a sub-class of Sub
that would handle operators.  And use this capability in the Raku
grammar to simplify a lot of actions.

This commit therefore:
- adds the Operator class in bootstrap and settings
- marks all pieces of code that can be removed once the Raku grammar
  is being used to build the setting
- adjust core classes tests to include Operator