RakuAST: create some raku interface methods for cursors
Instead of directly referring to NQP's private methods.  The idea
is that this will allow for easier debugging, documentation and
maybe open up possibilities for refinement / optimization.

More to come, but pushing for better bisectability now, as this is
very central to the Raku Programming Language and mistakes are
easily made.