Fix Backtrace view on what frames it considers "interesting"
What is coming from 'NQPCORE.setting' and 'BOOTSTRAP/v6?' is now
considered part of the setting. The former fixes uses of `NQPLock`.
The latter – frames from withing dispatchers, for example.

Don't stop building trace when any `NQP::` file is encountered. Relying
on the prefix alone was causing frames, following calls to
`NQPLock.protect` in the trace, to be skipped until (if we are lucky) a
`SETTING::` prefix found. This fix (and I hope this is a fix, not a
regression) makes the trace to skip only frames after method `eval` in
`NQPHLL/Compiler.nqp`.