RakuAST: Create find-nodes-exclusive
This creates an alternative implementation of 'find-nodes', but this
times it takes a full-on Code object as a `$stopper`. That `$stopper` is
called on each node and this check is designed to be _exclusive_, so
that nodes that match `$stopper` are never checked against the type
check and the `$condition`.