RakuAST: Migrate to WhateverApplicable
This allows a top-down approach to searching for other
`WhateverApplicable` nodes, building up a list of applicable child
nodes to curry across. This simplifies the actual implementation
of the "currying question" inside of the `Apply*` code significantly
while also covering the creation of deeply nested (or otherwise weird)
`WhateverCode` objects.

This isn't to imply that this implementation is "simple"... anything
as complex as the `*` in Raku is going to be a bit of a mind warp to
comprehend, let alone implement.