RakuAST: add ::L10N module
This module is intended to be used by any logic that wishes to create
code for use in localization (such as internal scripts, or external
developers, or just when developing other tools).

It exports two subroutines, each taking a language indicator, and a
hash with translation information, which return a RakuAST tree of a
role.

- "slangify" returns a role for mixing into the Raku grammar to activate
  localization.
- "deparsify" returns a role for mixing into the deparser, to activate
  deparsing in that localization.

The tools/build/makeL10N.raku is one of the users of this module.