RakuAST: generate localization files
This introduces the tools/build/makeL10N.raku helper script, to be run
everytime a change has been made in any of the translation files located
at tools/templates/L10N.

It also adds the initial incarnation of the L10N::NL module, which can
be used to activate the Dutch localization.

It also changes the already existing deparsing logic for the Dutch
localization, as it is now generated by the "makeL10N" script.