RakuAST: Migrate to more specific subclass
Instead of being merely an anonymous `Term::Named`, we
switch to using the much more descriptive
`WhateverCode::Argument`. It has been deliberately placed
outside of the `Term::*` hierarchy because it is almost
more of an implicit variable than it is a "term" in the
syntactic sense. But since it is the "curried"
(or "auto-primed") form of a term, it needs to maintain
the semantics of a term (particularly with regards to
binding to parameters).

This simplifies deparsing, which no longer needs to check
every `Var::Lexical` node to see whether the name starts
with `$whatevercode_arg_`. (Note: This otherwise would have
needed to be fixed to check every `Term::Named` instead).

While we are at it, we append a little information about the
`CurryThunk`: a psuedo-signature of its curried parameters.
I went with 🔆 because it resembles an asterix with a
hole in the middle: a whatever with a space in the center where
the eventually-provided argument goes.

We also change the `thunk-kind` of `CurryThunk` to the more
accurate `WhateverCode`. After all, it is the type you get back
from the AST expression once it goes through the rest of the
runtime. Also: we can technically be a mixture of
curried Whatevers and curried WhateverCodes, so `Curried Whatever`
is not properly descriptive.

`CurryThunk` reports its cardinality via the new `thunk-details`
feature, which allows curried nodes to add some details to their
`thunk-kind` reporting. Hopefully this proves useful for other
thunk kinds.

`WhateverCode::Argument` adds a 🔆 via `dump-markers`.