RakuAST: Implement FIRST phaser as an AST construct
It turns out that there is no need to even dip into QAST in order
to implement the `FIRST` phaser, which was particularly tricky
because it depended on so-called "extops" that are particularly
hard to arrange in pure QAST (`p6setfirstflag` and `p6takefirstflag`,
for the curious).

It turns out (nine++) that we can simply arrange for this construct
at an AST level, by emulating something like:

```
{
  state $triggered;
  unless $triggered {
    # ... the actual content of the FIRST phaser block
    $triggered = True
  }
}()
```

This feels like a much more natural implementation and it really
shows off what a difference RakuAST can already make to code generation.