RakuAST: Add some dump-markers
Currently AST is mainly decorated by details from any parsing stage
that occurred. This is fine except when AST is purely synthetic,
at which point very little is reported back to the user.

This is a first step towards any efforts to annotate more nodes with
`dunp-markers`.

【 and 】 were chosen because they give an impression of solidity
and what we are attempting to convey is the concrete *thing* that
distinguishes the node (usually a name, but perhaps sometimes
otherwise).

Oh, and we'll use 🛸 to indicate phasers, since they beam blorsts
up to their respective phases.