RakuAST: change interface if ::Postcircumfix::HashIndex.new
Only if there can only be one argument to an RakuAST::class.new,
should it be a single positional.  However, all of the ::Postcircumfix
classes will need to have a :colonpairs argument added to allow
in-place creation of these classes *with* adverbs.  So change the
interface of this class now, to not hide it later when the :colonpairs
logic gets added.