RakuAST: Add :{} object hash initialization
This turns out to be a simple case of enabling multiple identities
for `RakuAST::Circumfix::HashComposer` as both `&circumfix:<{ }>`
and `&circumfix:<:{ }>`.

Send the :object-hash reality down from the grammar through the
block-or-hash calculation into the composer itself, set the
resolver to use the right lookup in the object-hash initialization
case, and we're done.