Remove undocumented quotelang extenders
The primary quotelang markers q, qq, and Q allow for a number of extenders
according to the documentation: w, ww, and x.  So instead of qw, one could
also say q:w, and instead of qqww one could also say qq:ww.

The grammar however **ALSO** allows for: to, s, a, h, f, c, b, o.  Allowing
one to say qqto and Qf (to name but two of the possible combinations).

I think this is taking the TIMTOWTDO a little too far.  So this commit
removes support for these undocumented quotelang extenders.

This breaks three spectest files:
- t/spec/S02-literals/misc-interpolation.t
- t/spec/S02-literals/quoting.t
- t/spec/integration/advent2014-day16.t

because they check for a, b and to as quotelang extenders respectively.
If this change is OKd, then these tests will have to be adapted.  The
adaptation is quite simple: add a colon.