RakuAST: add localization of quote-lang markers
Specifically: m ms q Q qq rx s S ss Ss .

Technically, ms ss Ss are really synonyms for m:ss s:ss S:ss respectively
but that proved very hard to solve with the given grammar tokens,
because all of these markers need to check whether there is an entity
defined with the same name, which as such changes the parsing.

Example:
    use Test;
    my \m = 42;
    is m, 42, 'the answer';

Note that in the above example, the "m" is **not** parsed as a
quote-lang marker, but as a term, even though it is followed by a
delimiter.