Fix handling of dynamic containers on 6.e PseudoStash
Before it was trying to provide support for Array and Hash containers.
But since `.VAR` returns the original object when there is no `Scalar`
container the code was considering as dynamic symbols any kind of symbol
which has a returning true `dynamic` method. Aside of that, attempt to
test for the method on roles was causing them to pun unintentionally.