RakuAST: make sure localizations work for foo()
This was taking an alternate path to create Call::Name object, and
thus missing the core2ast translation step.  And this was wreaking
havoc on serialization/deserialization tests of localizations.

Problem was that the alternate path did not actually have an .ast
connected to the match object yet.  After adaptation of the default
case (in e008984a87) this changes the generated code for the xx2ast
methods to allow match objects without .ast and just generate a
RakuAST::Name object for the stringification of the match object.