RakuAST: fix deparsing of attributes with initializers
These appear to be codegenned with both an initializer value, as
well as a "will build" trait.  Not sure whether that is always needed
or not.  Ignore any "will build" traits in deparsing for now.

Also add preliminary rakufication for RakuAST::Trait::Will objects
so that attributes with initializers will at least rakufy without
bombing.