More sensible exception on wrong type in attribute's `is default`
There is a contradiction in S02-names/is_default.t where TODO-ed tests
expect `X::Parameter::Default::TypeCheck`, but a few active ones are
checking for `X::TypeCheck::Assignment`.

The problem with `X::Parameter::Default::TypeCheck` is that it is
neither `X::TypeCheck`, nor it provides correct information about the
problem because it is aimed at parameter binding in first place.

The problem with `X::TypeCheck::Asignment` is that applying `default`
trait isn't an assignment. Consequently, the error message is slightly
misleading.

Exception `X::TypeCheck::Attribute::Default` speaks for itself and using
somewhat better wording in the error message as it mentions "default
value". The problem with this solution is that specs needs to be updated
too.