RakUAST: introduce RakuAST::Package::Attachable
Not all -package- types can have methods or attributes attached.  This
commit moves all logic with regards to attaching method and attributes
from the RakuAST::Package class to the RakuAST::Package::Attachable
class, and makes the RakuAST::Role/Class classes subclass from that.

This allows simplification of the RakuAST::Package class, and allows
an external test to see if a package can have attributes/methods or
not (by checking for RakuAST::Package::Attachable).

RakuAST::Role|Class now have a subclassed PRODUCE-META-OBJECT method
that takes care of their specific peculiarities, rather than having
to check for that in the generic PRODUCE-META-OBJECT method.